requirements.txt
streamlit
requests
pandas
feedparser
gspread
google-auth
google-auth-oauthlib
